+++ src/Metrics/DTO/Values/MetricValueCollection.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\Metrics\DTO\Values;
+
+/**
+ * Class ValueCollection
+ *
+ * @method attach(AbstractValue $metricValue)
+ * @method current(): AbstractValue $metricValue
+ *
+ * @package B24io\Loyalty\SDK\Metrics\DTO\Values
+ */
+class MetricValueCollection extends \SplObjectStorage
+{
+}