+++ src/Metrics/Exceptions/InvalidMetricCodeException.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\Metrics\Exceptions;
+
+use B24io\Loyalty\SDK\Exceptions\ObjectInitializationException;
+
+/**
+ * Class InvalidMetricCodeException
+ *
+ * @package B24io\Loyalty\SDK\Metrics\DTO
+ */
+class InvalidMetricCodeException extends ObjectInitializationException
+{
+}