+++ src/Metrics/Transport/Admin/Transport.php
@@ -29,8 +29,12 @@ class Transport extends SDK\Transport\AbstractTransport
 
         $requestResult = $this->apiClient->executeApiRequest(
             sprintf(
-                'admin/metrics/?loyalty_client_api_key=%s',
-                $this->apiClient->getAuthToken()->getClientApiKey()
+                'admin/metrics/?%s',
+                http_build_query(
+                    [
+                        self::CLIENT_API_KEY => $this->apiClient->getAuthToken()->getClientApiKey()->toString(),
+                    ]
+                )
             ),
             RequestMethodInterface::METHOD_GET
         );
@@ -48,5 +52,60 @@ class Transport extends SDK\Transport\AbstractTransport
         );
 
         return $metricCollectionResponse;
+    }
+
+    /**
+     * @param Metrics\DTO\MetricCode $metricCode
+     * @param \DateTime              $dateFrom
+     * @param \DateTime              $dateTo
+     *
+     * @return Metrics\Transport\DTO\MetricReportResponse
+     * @throws SDK\Exceptions\ApiClientException
+     * @throws SDK\Exceptions\NetworkException
+     * @throws SDK\Exceptions\TransportFormatException
+     * @throws SDK\Exceptions\UnknownException
+     */
+    public function getReportByMetricCode(
+        Metrics\DTO\MetricCode $metricCode,
+        \DateTime $dateFrom,
+        \DateTime $dateTo
+    ): Metrics\Transport\DTO\MetricReportResponse {
+        $this->log->debug(
+            'b24io.loyalty.sdk.Metrics.transport.admin.getReportByMetricCode.start',
+            [
+                'metricCode' => $metricCode->getValue(),
+                'dateFrom'   => $dateFrom->format(\DATE_ATOM),
+                'dateTo'     => $dateTo->format(\DATE_ATOM),
+            ]
+        );
+
+        $requestResult = $this->apiClient->executeApiRequest(
+            sprintf(
+                'admin/metrics/%s/?%s',
+                $metricCode->getValue(),
+                http_build_query(
+                    [
+                        self::CLIENT_API_KEY => $this->apiClient->getAuthToken()->getClientApiKey(),
+                        self::DATE_FROM      => $dateFrom->format(\DATE_RFC3339),
+                        self::DATE_TO        => $dateTo->format(\DATE_RFC3339),
+                    ]
+                )
+            ),
+            RequestMethodInterface::METHOD_GET
+        );
+
+        $metricReportResponse = new SDK\Metrics\Transport\DTO\MetricReportResponse(
+            SDK\Metrics\DTO\Fabric::initReportFromArray($requestResult['result']),
+            $this->initMetadata($requestResult['meta'])
+        );
+
+        $this->log->debug(
+            'b24io.loyalty.sdk.Metrics.transport.admin.getReportByMetricCode.finish',
+            [
+                'metadata' => SDK\Transport\Formatters\Metadata::toArray($metricReportResponse->getMeta()),
+            ]
+        );
+
+        return $metricReportResponse;
     }
 }