+++ src/OperationsJournal/Transport/User/Transport.php
@@ -40,10 +40,14 @@ class Transport extends SDK\Transport\AbstractTransport
 
         $requestResult = $this->apiClient->executeApiRequest(
             sprintf(
-                'user/operations-journal/get-by-card-uuid/%s/?dateFrom=%s&dateTo=%s',
+                'user/operations-journal/get-by-card-uuid/%s/?%s',
                 $this->apiClient->getAuthToken()->getAuthApiKey()->toString(),
-                urlencode($dateFrom->format(\DATE_RFC3339)),
-                urlencode($dateTo->format(\DATE_RFC3339))
+                http_build_query(
+                    [
+                        self::DATE_FROM => $dateFrom->format(\DATE_RFC3339),
+                        self::DATE_TO   => $dateTo->format(\DATE_RFC3339),
+                    ]
+                )
             ),
             RequestMethodInterface::METHOD_GET,
             []