+++ src/SystemJournal/Transport/Admin/Transport.php
@@ -2,49 +2,53 @@
 
 declare(strict_types=1);
 
-namespace B24io\Loyalty\SDK\OperationsJournal\Transport\Admin;
+namespace B24io\Loyalty\SDK\SystemJournal\Transport\Admin;
 
 use B24io\Loyalty\SDK;
-use B24io\Loyalty\SDK\Cards;
+use B24io\Loyalty\SDK\SystemJournal;
 
 use Fig\Http\Message\RequestMethodInterface;
 
 /**
  * Class Transport
  *
- * @package B24io\Loyalty\SDK\Turnover\Transport\Admin
+ * @package B24io\Loyalty\SDK\SystemJournal\Transport\Admin
  */
 class Transport extends SDK\Transport\AbstractTransport
 {
     /**
-     * @param \DateTime $dateFrom
-     * @param \DateTime $dateTo
+     * @param \DateTime                       $dateFrom
+     * @param \DateTime                       $dateTo
+     * @param SystemJournal\DTO\LogLevel|null $level
      *
-     * @return SDK\OperationsJournal\Transport\DTO\OperationsJournalResponse
+     * @return SystemJournal\Transport\DTO\SystemJournalResponse
      * @throws SDK\Exceptions\ApiClientException
      * @throws SDK\Exceptions\NetworkException
      * @throws SDK\Exceptions\TransportFormatException
      * @throws SDK\Exceptions\UnknownException
      */
-    public function getOperationsByPeriod(
+    public function getSystemJournalByPeriod(
         \DateTime $dateFrom,
-        \DateTime $dateTo
-    ): SDK\OperationsJournal\Transport\DTO\OperationsJournalResponse {
+        \DateTime $dateTo,
+        ?SystemJournal\DTO\LogLevel $level
+    ): SystemJournal\Transport\DTO\SystemJournalResponse {
         $this->log->debug(
-            'b24io.loyalty.sdk.OperationsJournal.transport.admin.getOperationsByPeriod.start',
+            'b24io.loyalty.sdk.SystemJournal.transport.admin.getSystemJournalByPeriod.start',
             [
                 'dateFrom' => $dateFrom->format(\DATE_ATOM),
                 'dateTo'   => $dateTo->format(\DATE_ATOM),
+                'level'    => $level !== null ? $level->key() : '',
             ]
         );
 
         $requestResult = $this->apiClient->executeApiRequest(
             sprintf(
-                'admin/operations-journal/?%s',
+                'admin/system-journal/?%s',
                 http_build_query(
                     [
                         self::DATE_FROM => $dateFrom->format(\DATE_RFC3339),
                         self::DATE_TO   => $dateTo->format(\DATE_RFC3339),
+                        'level'         => $level !== null ? $level->key() : '',
                     ]
                 )
             ),
@@ -52,13 +56,13 @@ class Transport extends SDK\Transport\AbstractTransport
             []
         );
 
-        $operationsJournal = new SDK\OperationsJournal\Transport\DTO\OperationsJournalResponse(
-            SDK\OperationsJournal\DTO\Fabric::initOperationsJournalFromArray($requestResult['result'], $this->log),
+        $operationsJournal = new SDK\SystemJournal\Transport\DTO\SystemJournalResponse(
+            SDK\SystemJournal\DTO\Fabric::initSystemJournalFromArray($requestResult['result']),
             $this->initMetadata($requestResult['meta'])
         );
 
         $this->log->debug(
-            'b24io.loyalty.sdk.turnover.transport.admin.getOperationsByPeriod.finish',
+            'b24io.loyalty.sdk.SystemJournal.transport.admin.getSystemJournalByPeriod.finish',
             [
                 'metadata' => SDK\Transport\Formatters\Metadata::toArray($operationsJournal->getMeta()),
             ]